Working on CurvePalette
In the making still, there's some bugs to solve. Holding the alt-key allows to make a selection marquee in the CurvePalette grid view to select points that are far our of the normal gray diagonal.
Besides the grid colors, the points are also show on their unrounded "tension" positions. A darker outline shows that they are selected.