Similar spacing
Spacer sample (•) Similarity mode now shows the glyphs for left & right that have a similar shape.
Click on the glyphs the make it current.
(Known bug still, the selection on the spacer/kerning is still a bit off to the right)
Next is to copy side bearings of the base glyphs to the glyphs with similar sides.